chore: update spec-update.testbed-tests
