+++ tests/spec/test-data/spec-update.testbed-tests/c.ts
@@ -1,0 +1,8 @@
+import { LogService, BDep } from '@angular/core';
+
+export class C  {
+    constructor(
+        private bDep: bDep,
+        private logger: LogService
+    ) {}
+}
